+++ navigation/Stack.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import Signup from '../screens/Signup';
+import Signup from '../screens/signup/Signup';
 
 const NativeStack = createNativeStackNavigator();
